readme.txt
This is one of the problems from geektrust.com i have solved and the source code does not contain any third party library or even an imported python library. This code can easily be executed on any python 3.x machine with the test cases in the same format as specified in the problem.